SPES-server/requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
cffi==1.15.0
click==8.0.4
cryptography==36.0.1
dependency-injector==4.38.0
Deprecated==1.2.13
fastapi==0.74.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
numpy==1.22.3
packaging==21.3
passlib==1.7.4
psycopg2-binary==2.9.3
pycountry==22.3.5
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.7
python-multipart==0.0.5
PyYAML==6.0
redis==4.2.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
typing_extensions==4.1.1
uvicorn==0.17.5
wrapt==1.14.0
